Fixed issues with CoreEx with linknodes; Added fox news data
CoreEx now only determines if its parent node is a link node. Doing so seemed to grab more content than before. The or operator that was used in the conditional before was maybe scoring links higher than it should have. Thus a lot of links were pulled in from sites like CNN and foxnews that have a ton of extra junk, lol. BBC seems to perform the best as far as extracting content so far. Also added Nicks material from fox news to the data set