Finished fine tuning the three different models;  Performance folder
For evaluating link target Id using C4.5 in addition to ten way cross-validation and pruning,  the grouping of attribute values and windowing will be used(ie ten trees will be grown and the best tree chosen). For evaluating CoreEx using C4.5, only the grouping of attribute values will be used in addition to 10 way cross-validation and pruning. For evaluating both CoreEx and link-target Id, only windowing will be used in addition to 10 way cross-validation and pruning. I also added a performance folder for each model in which the results of each trial conducted (namely, 20) will be stored.